Add overlay configuration functionality
Allow a user to overwrite just specific configuration files on container start. Introduced the etc-overlay directory
convention. All files overwrite the default configuration on container start. Introduced the convention to give users
more control to do upgrades by adding a "do-upgrade" file in the etc-overlay directory.

Health checks are now completely on service stack description level and not in the container itself gives more
flexibility deploying the containers. It is now required to use a current Docker and Docker-Compose version which
supports those health checks. The README has updated accordingly.

Removed the Horizon container delay to wait on PostgreSQL database by using docker compose dependency condition
"healthy" for the PostgreSQL database.

Migrated the container build workflow from DockerHub to CircleCi. The images are still hosted and pushed to DockerHub
and are not changed.